add person works as expected

+++ models/setup.sql
@@ -4,6 +4,6 @@ CREATE TABLE person(
     id serial PRIMARY KEY,
     uuid VARCHAR(255),
     name VARCHAR(255),
-    age INTEGER(11),
+    age INTEGER,
     created_at timestamp not null
 );
